more reuse on sensor box initialization code
