+++ CA_DataUploaderLib/BaseSensorBox.cs
@@ -2,7 +2,6 @@
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading;
-using System.Globalization;
 using System.Text;
 using CA_DataUploaderLib.IOconf;
 using System.Text.RegularExpressions;
@@ -22,8 +21,31 @@ namespace CA_DataUploaderLib
         protected List<SensorSample> _values = new List<SensorSample>();
         protected List<MCUBoard> _boards = new List<MCUBoard>();
         protected int expectedHeaderLines = 8;
+        private string commandHelp;
 
-        public BaseSensorBox() { }
+        public BaseSensorBox(CommandHandler cmd, string commandName, string commandArgsHelp, string commandDescription, IEnumerable<IOconfInput> values) 
+        { 
+            Title = commandName;
+            _cmd = cmd;
+            _logLevel = IOconfFile.GetOutputLevel();
+            _values = values.IsInitialized().Select(x => new SensorSample(x)).ToList();
+            if (!_values.Any())
+                return;  // no data
+
+            if (cmd != null)
+            {
+                commandHelp = $"{commandName.ToLower() + " " + commandArgsHelp,-22}- {commandDescription}";
+                cmd.AddCommand(commandName.ToLower(), ShowQueue);
+                cmd.AddCommand("help", HelpMenu);
+                cmd.AddCommand("escape", Stop);
+            }
+
+            _boards = _values.Where(x => !x.Input.Skip).Select(x => x.Input.Map.Board).Distinct().ToList();
+            CALog.LogInfoAndConsoleLn(LogID.A, $"{commandName} boards: {_boards.Count()} values: {_values.Count()}");
+
+            _running = true;
+            new Thread(() => LoopForever()).Start();
+        }
 
         public SensorSample GetValueByTitle(string title) =>
                 _values.SingleOrDefault(x => x.Input.Name == title) ??
@@ -133,6 +155,7 @@ namespace CA_DataUploaderLib
             }
         }
 
+        // TODO: allow to override the parsing so that we can parse the ScaleBox line that has kg at the end.
         /// <returns>the list of doubles, otherwise <c>null</c></returns>
         protected static List<double> TryParseAsDoubleList(string row)
         {
@@ -217,6 +240,12 @@ namespace CA_DataUploaderLib
         protected int GetHubID(SensorSample sensor)
         {
             return _values.GroupBy(x => x.Input.BoxName).Select(x => x.Key).ToList().IndexOf(sensor.Input.BoxName);
+        }
+
+        private bool HelpMenu(List<string> args)
+        {
+            CALog.LogInfoAndConsoleLn(LogID.A, $"flows                 - show all flows in input queue");
+            return true;
         }
 
         public void Dispose()
