+++ CA_DataUploaderLib/ThermocoupleBox.cs
@@ -2,7 +2,6 @@
 using CA_DataUploaderLib.Extensions;
 using System.Collections.Generic;
 using System.Linq;
-using System.Threading;
 using CA_DataUploaderLib.Helpers;
 using System;
 
@@ -12,47 +11,15 @@ namespace CA_DataUploaderLib
     {
         private readonly SensorSample _rpiGpuSample;
         private readonly SensorSample _rpiCpuSample;
-        public ThermocoupleBox(CommandHandler cmd)
+        public ThermocoupleBox(CommandHandler cmd) : base(cmd, "Temperatures", string.Empty, "show all temperatures in input queue", IOconfFile.GetTypeKAndLeakage()) 
         {
-            Title = "Thermocouples";
-            _cmd = cmd;
-            _logLevel = IOconfFile.GetOutputLevel();
-
-            _values = IOconfFile.GetTypeKAndLeakage().IsInitialized().Select(x => new SensorSample(x)).ToList();
-            var rpiTemp = IOconfFile.GetRPiTemp();
-            if (!rpiTemp.Disabled && !OperatingSystem.IsWindows())
-            {
-                _values.Add(_rpiGpuSample = new SensorSample(rpiTemp.WithName(rpiTemp.Name + "Gpu")));
-                _values.Add(_rpiCpuSample = new SensorSample(rpiTemp.WithName(rpiTemp.Name + "Cpu")));
-            }
-
-            if (!_values.Any())
-                return;
-
-            if (cmd != null)
-            {
-                cmd.AddCommand("temperatures", ShowQueue);
-                cmd.AddCommand("help", HelpMenu);
-                cmd.AddCommand("Junction", Junction);
-                cmd.AddCommand("escape", Stop);
-            }
-
-            _boards = _values.Where(x => !x.Input.Skip).Select(x => x.Input.Map.Board).Distinct().ToList();
-            CALog.LogInfoAndConsoleLn(LogID.A, $"ThermocoupleBox boards: {_boards.Count()} values: {_values.Count()}");
-
-            foreach (var board in _boards)
-            {
-                if(_values.Any(x=> x.Input.PortNumber == 0 && x.Input.BoxName == board.BoxName && ((IOconfTypeK)x.Input).AllJunction))
-                    board.WriteLine("Junction");
-            }
-
-            new Thread(() => this.LoopForever()).Start();
+            _rpiGpuSample = _values.FirstOrDefault(x => x.Name.EndsWith("Gpu"));
+            _rpiCpuSample = _values.FirstOrDefault(x => x.Name.EndsWith("Cpu"));
         }
 
         protected override void ReadSensors()
         {
             base.ReadSensors();
-
 
             if (_rpiGpuSample != null)
                 _rpiGpuSample.Value = DULutil.ExecuteShellCommand("vcgencmd measure_temp").Replace("temp=", "").Replace("'C", "").ToDouble();
@@ -60,18 +27,12 @@ namespace CA_DataUploaderLib
                 _rpiCpuSample.Value = DULutil.ExecuteShellCommand("cat /sys/class/thermal/thermal_zone0/temp").ToDouble() / 1000;
         }
 
-        private bool HelpMenu(List<string> args)
+        private IEnumerable<IOconfInput> GetSensors()
         {
-            CALog.LogInfoAndConsoleLn(LogID.A, $"temperatures              - show all temperatures in input queue");
-            return true;
-        }
-
-        private bool Junction(List<string> args)
-        {
-            foreach (var board in _boards)
-                board.WriteLine("Junction");
-
-            return true;
+            var values = IOconfFile.GetTypeKAndLeakage();
+            var rpiTemp = IOconfFile.GetRPiTemp();
+            var addRpiTemp = !rpiTemp.Disabled && !OperatingSystem.IsWindows();
+            return addRpiTemp ? values.Concat(new []{ rpiTemp.WithName(rpiTemp.Name + "Gpu"), rpiTemp.WithName(rpiTemp.Name + "Cpu")}) : values;
         }
     }
 }
